double x,y became int x,y and all related methods adjusted according to that. a.4 is done.

+++ src/Locatable.java
@@ -1,5 +1,5 @@
 public interface Locatable {
     double getX();
     double getY();
-    void setLocation(double x, double y);
+    void setLocation(int x, int y);
 }
